backend/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
pydantic
aiofiles
biopython
python-multipart
reportlab
google-generativeai
numpy
biopandas
matplotlib 
prody
pandas
py3dmol
rdkit
plotly